requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cffi==1.16.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
typing_extensions==4.9.0
wavelink==3.1.2
yarl==1.9.2
